requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
# Dependencias para 2FA
pyotp==2.9.0
qrcode[pil]==8.2
pillow==12.0.0
